🐛🧪🎨 Make lexer test compare tokens
With new releases of `pygments`, this test was getting broken again
and again because it originally relied on HTML the formatted text
output. Hence, it was prone to external influence for no good reason.

This patch rewrites it to solely rely on the underlying tokens that
our lexer produces, therefore reducing the dependence on unrelated
changes in the `pygments` library.